requirements.txt
antlr4-python3-runtime==4.9.3
opencv-python
pyrealsense2
appdirs==1.4.4
brotlipy==0.7.0
certifi
cffi
charset-normalizer
click
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpickle
contourpy
cryptography
cycler
cytoolz
dask
filelock==3.12.4
fonttools==4.25.0
fsspec
ftfy==6.1.1
future
huggingface-hub==0.0.12
hydra-core==1.3.2
idna
imagecodecs
imageio
importlib-metadata
inplace-abn==1.1.0
joblib
kiwisolver
locket
matplotlib
munkres==1.1.4
networkx
numpy
omegaconf==2.3.0
packaging
partd
Pillow==9.4.0
piq==0.8.0
ply==3.11
pooch
protobuf==3.20.3
pycparser
pyOpenSSL
pyparsing
PyQt5-sip==12.11.0
PySocks
python-dateutil
PyWavelets
PyYAML
regex==2023.8.8
requests
ruamel.yaml
ruamel.yaml.clib
sacremoses==0.0.53
safetensors==0.3.3
scikit-image
scipy==1.10.1
sip
six
tensorboardX
tifffile
timm==0.9.7
tokenizers==0.10.3
toml
toolz
torch
torchsummary==1.5.1
torchvision
tornado
tqdm==4.66.1
transformers==4.8.1
typing_extensions
urllib3
wcwidth==0.2.6
zipp
